implementando a impressão de moedas e notas

+++ src/TrocoMoeda.java
@@ -1,10 +1,10 @@
 
 public class TrocoMoeda {
 	
-        double moeda;
-        double quantidade;
+        int moeda;
+        int quantidade;
 
-        public TrocoMoeda(double moeda, double quantidade) {
+        public TrocoMoeda(int moeda, int quantidade) {
             this.moeda = moeda;
             this.quantidade = quantidade;
         }
